Much better results, but only half the eigenvalues agree.
This seems to be a problem with the degeneracy; one of the eigenvalues is the degenerate pair is good, the other is terrible. The precision is already high, so I'm not sure what to do, other than take the matrix directly from ReSpect, and see if the error really is due to slight differences in the submatrix, or is infact and issue with the eigensolver.